Added method getModules Cleanup of files
