+++ lib/Contracts.js
@@ -32,9 +32,8 @@ class OpenSTContracts extends Contracts {
    * @returns {auxiliaryWeb3.eth.Contract}
    */
   TokenRules(contractAddress, txOptions) {
-    const oThis = this;
     const jsonInterface = abiBinProvider.getABI('TokenRules');
-    const contractInstance = new oThis.auxiliaryWeb3.eth.Contract(jsonInterface, contractAddress, txOptions);
+    const contractInstance = new this.auxiliaryWeb3.eth.Contract(jsonInterface, contractAddress, txOptions);
     return contractInstance;
   }
 
@@ -47,9 +46,8 @@ class OpenSTContracts extends Contracts {
    * @returns {auxiliaryWeb3.eth.Contract}
    */
   PricerRule(contractAddress, txOptions) {
-    const oThis = this;
     const jsonInterface = abiBinProvider.getABI('PricerRule');
-    const contractInstance = new oThis.auxiliaryWeb3.eth.Contract(jsonInterface, contractAddress, txOptions);
+    const contractInstance = new this.auxiliaryWeb3.eth.Contract(jsonInterface, contractAddress, txOptions);
     return contractInstance;
   }
 
@@ -62,10 +60,8 @@ class OpenSTContracts extends Contracts {
    * @returns {auxiliaryWeb3.eth.Contract}
    */
   GnosisSafe(contractAddress, txOptions) {
-    const oThis = this;
-
     const jsonInterface = abiBinProvider.getABI('GnosisSafe');
-    let contractInstance = new oThis.auxiliaryWeb3.eth.Contract(jsonInterface, contractAddress, txOptions);
+    let contractInstance = new this.auxiliaryWeb3.eth.Contract(jsonInterface, contractAddress, txOptions);
     return contractInstance;
   }
 
@@ -78,10 +74,8 @@ class OpenSTContracts extends Contracts {
    * @returns {auxiliaryWeb3.eth.Contract}
    */
   TokenHolder(contractAddress, txOptions) {
-    const oThis = this;
-
     const jsonInterface = abiBinProvider.getABI('TokenHolder');
-    let contractInstance = new oThis.auxiliaryWeb3.eth.Contract(jsonInterface, contractAddress, txOptions);
+    let contractInstance = new this.auxiliaryWeb3.eth.Contract(jsonInterface, contractAddress, txOptions);
     return contractInstance;
   }
 
@@ -95,10 +89,9 @@ class OpenSTContracts extends Contracts {
    * @constructor
    */
   static getTokenRules(auxiliaryWeb3, address, options) {
-    const oThis = this;
     const jsonInterface = abiBinProvider.getABI('TokenRules');
-    const contract = new auxiliaryWeb3.eth.Contract(jsonInterface, address, options);
-    return contract;
+    const contractInstance = new auxiliaryWeb3.eth.Contract(jsonInterface, address, options);
+    return contractInstance;
   }
 
   /**
@@ -111,10 +104,9 @@ class OpenSTContracts extends Contracts {
    * @constructor
    */
   static getPricerRule(auxiliaryWeb3, address, options) {
-    const oThis = this;
     const jsonInterface = abiBinProvider.getABI('PricerRule');
-    const contract = new auxiliaryWeb3.eth.Contract(jsonInterface, address, options);
-    return contract;
+    const contractInstance = new auxiliaryWeb3.eth.Contract(jsonInterface, address, options);
+    return contractInstance;
   }
 
   /**
@@ -127,10 +119,9 @@ class OpenSTContracts extends Contracts {
    * @constructor
    */
   static getTokenHolder(auxiliaryWeb3, address, options) {
-    const oThis = this;
     const jsonInterface = abiBinProvider.getABI('TokenHolder');
-    const contract = new auxiliaryWeb3.eth.Contract(jsonInterface, address, options);
-    return contract;
+    const contractInstance = new auxiliaryWeb3.eth.Contract(jsonInterface, address, options);
+    return contractInstance;
   }
 
   /**
@@ -143,10 +134,9 @@ class OpenSTContracts extends Contracts {
    * @constructor
    */
   static getGnosisSafe(auxiliaryWeb3, address, options) {
-    const oThis = this;
     const jsonInterface = abiBinProvider.getABI('GnosisSafe');
-    const contract = new auxiliaryWeb3.eth.Contract(jsonInterface, address, options);
-    return contract;
+    const contractInstance = new auxiliaryWeb3.eth.Contract(jsonInterface, address, options);
+    return contractInstance;
   }
 
   /**
@@ -158,7 +148,6 @@ class OpenSTContracts extends Contracts {
    * @returns {auxiliaryWeb3Object.eth.Contract}
    */
   static getDelayedRecovery(auxiliaryWeb3, address, options) {
-    const oThis = this;
     const jsonInterface = abiBinProvider.getABI('DelayedRecoveryModule');
     const contractInstance = new auxiliaryWeb3.eth.Contract(jsonInterface, address, options);
     return contractInstance;
