+++ lib/helper/GnosisSafe.js
@@ -365,6 +365,20 @@ class GnosisSafe {
 
     return contract.methods.domainSeparator().call();
   }
+
+  /**
+   * Returns All the modules stored in gnosis contract.
+   *
+   * @returns {Promise<*>}
+   */
+  async getModules() {
+    const oThis = this;
+
+    const jsonInterface = oThis.abiBinProvider.getABI(gnosisSafeContractName);
+    const contract = new oThis.auxiliaryWeb3.eth.Contract(jsonInterface, oThis.gnosisSafeProxy);
+
+    return contract.methods.getModules().call();
+  }
 }
 
 module.exports = GnosisSafe;
