+++ test/unit/recovery/executeRecovery.js
@@ -4,12 +4,12 @@ const Web3 = require('web3');
 const sinon = require('sinon');
 const { assert } = require('chai');
 
-const Spy = require('../../utils/Spy');
-const Recovery = require('../../lib/ContractInteract/Recovery');
-const AssertAsync = require('../../utils/AssertAsync');
-const Utils = require('../../utils/Utils');
+const Spy = require('../../../utils/Spy');
+const Recovery = require('../../../lib/ContractInteract/Recovery');
+const AssertAsync = require('../../../utils/AssertAsync');
+const Utils = require('../../../utils/Utils');
 
-describe('Recovery.abortRecoveryByOwner()', () => {
+describe('Recovery.executeRecovery()', () => {
   let recovery;
   let web3;
 
@@ -23,22 +23,19 @@ describe('Recovery.abortRecoveryByOwner()', () => {
     const prevOwner = '0x0000000000000000000000000000000000000003';
     const oldOwner = '0x0000000000000000000000000000000000000004';
     const newOwner = '0x0000000000000000000000000000000000000005';
-    const r = 'r';
-    const v = 'v';
-    const s = 's';
 
     const mockRawTx = 'mockRawTx';
 
-    const rawTx = sinon.replace(recovery, 'abortRecoveryByOwnerRawTx', sinon.fake.resolves(mockRawTx));
+    const rawTx = sinon.replace(recovery, 'executeRecoveryRawTx', sinon.fake.resolves(mockRawTx));
 
     const spySendTransaction = sinon.replace(Utils, 'sendTransaction', sinon.fake.resolves(true));
     const txOptions = {
       from: '0x0000000000000000000000000000000000000006'
     };
 
-    const response = await recovery.abortRecoveryByOwner(prevOwner, oldOwner, newOwner, r, s, v, txOptions);
-    assert.isTrue(response, 'abortRecoveryByOwner should return true');
-    Spy.assert(rawTx, 1, [[prevOwner, oldOwner, newOwner, r, s, v]]);
+    const response = await recovery.executeRecovery(prevOwner, oldOwner, newOwner, txOptions);
+    assert.isTrue(response, 'executeRecovery should return true');
+    Spy.assert(rawTx, 1, [[prevOwner, oldOwner, newOwner]]);
     Spy.assert(spySendTransaction, 1, [[mockRawTx, txOptions]]);
     sinon.restore();
   });
@@ -47,13 +44,10 @@ describe('Recovery.abortRecoveryByOwner()', () => {
     const prevOwner = '0x0000000000000000000000000000000000000003';
     const oldOwner = '0x0000000000000000000000000000000000000004';
     const newOwner = '0x0000000000000000000000000000000000000005';
-    const r = 'r';
-    const v = 'v';
-    const s = 's';
     const txOptions = undefined;
 
     await AssertAsync.reject(
-      recovery.abortRecoveryByOwner(prevOwner, oldOwner, newOwner, r, s, v, txOptions),
+      recovery.executeRecovery(prevOwner, oldOwner, newOwner, txOptions),
       'Invalid transaction options: undefined.'
     );
   });
@@ -62,12 +56,9 @@ describe('Recovery.abortRecoveryByOwner()', () => {
     const prevOwner = '0x0000000000000000000000000000000000000003';
     const oldOwner = '0x0000000000000000000000000000000000000004';
     const newOwner = '0x0000000000000000000000000000000000000005';
-    const r = 'r';
-    const v = 'v';
-    const s = 's';
     const txOptions = {};
     await AssertAsync.reject(
-      recovery.abortRecoveryByOwner(prevOwner, oldOwner, newOwner, r, s, v, txOptions),
+      recovery.executeRecovery(prevOwner, oldOwner, newOwner, txOptions),
       `Invalid from address ${txOptions.from} in transaction options.`
     );
   });
@@ -76,14 +67,11 @@ describe('Recovery.abortRecoveryByOwner()', () => {
     const prevOwner = '0x0000000000000000000000000000000000000003';
     const oldOwner = '0x0000000000000000000000000000000000000004';
     const newOwner = '0x0000000000000000000000000000000000000005';
-    const r = 'r';
-    const v = 'v';
-    const s = 's';
     const txOptions = {
       from: '0x123'
     };
     await AssertAsync.reject(
-      recovery.abortRecoveryByOwner(prevOwner, oldOwner, newOwner, r, s, v, txOptions),
+      recovery.executeRecovery(prevOwner, oldOwner, newOwner, txOptions),
       `Invalid from address ${txOptions.from} in transaction options.`
     );
   });
