+++ test/unit/recovery/resetRecoveryOwner.js
@@ -4,10 +4,10 @@ const Web3 = require('web3');
 const sinon = require('sinon');
 const { assert } = require('chai');
 
-const Spy = require('../../utils/Spy');
-const Recovery = require('../../lib/ContractInteract/Recovery');
-const AssertAsync = require('../../utils/AssertAsync');
-const Utils = require('../../utils/Utils');
+const Spy = require('../../../utils/Spy');
+const Recovery = require('../../../lib/ContractInteract/Recovery');
+const AssertAsync = require('../../../utils/AssertAsync');
+const Utils = require('../../../utils/Utils');
 
 describe('Recovery.resetRecoveryByOwner()', () => {
   let recovery;
